pay: Add strict_first option to send_payment_core
send_payment_core uses non-strict channel selection on the first
hop. The newly introduced option strict_first forces the payment
core to use the channel with the scid given in the route for the
first hop.
If the channel can not add an htlc, the sendpay core returns an
error PAY_TRY_OTHER_ROUTE.